requirements.txt
anthropic>=0.40.0
PyQt6>=6.6.0
pyautogui>=0.9.54
mss>=9.0.0
pynput>=1.7.6
Pillow>=10.0.0
numpy>=1.24.0
